requirements.txt
pip>=23.0
pandas>=2.0.0
numpy>=1.24.0
scikit-fuzzy>=0.4.2
openpyxl>=3.1.0
nltk>=3.8.1
spacy>=3.7.0
scikit-learn>=1.3.0
fuzzywuzzy[speedup]>=0.18.0
python-levenshtein>=0.21.0
sentence-transformers>=2.6.1
torch>=2.1.0
Pillow>=10.0.0
opencv-python>=4.8.0
pdf2image>=1.16.3
streamlit>=1.30.0
plotly>=5.18.0
matplotlib>=3.8.0
easyocr
opencv-python-headless
